do not leak cached metadata on non-cached packages to the consolidated report
In some cases, namely when a utility (e.g. tests) did not run
for a package, we do not use the cached data. The cache_pull logic
will still fill the known metadata in the cache-pull report
for potential future use, but with 'cached' set to false.

This was not handled within the `consolidated_report`. In some cases,
we could end up with old metadata 'leaking' into the report, and with
non 'cached' entry. This commit explicitely removes this data from
the final report.