cli: implement dpkg-filter-status
The goal of the command is to create a list of "runtime" packages.
This would have been better handled by autoproj but that would be
a huge investment in time (both implementation and actually updating
the packages). We need a patch solution until we can invest in this
'proper' solution.