fix: update cache entries mtime during pull
I became convinced that rarely-updated but always-used cache entries
were being evicted in a 17G cache (so, unlikely to be actually
evicted by new useful stuff).

It turns out that the place where the cache entries mtime was not
enough. push_package_to_cache is not called for entries that were
pulled from the cache.

Touch the cache entries in the pull path, which is simpler and
guarantees that we won't have this happen again.