fix: explicitly cache log directories
CI currently relies on cached builds to contain their log files, so
that build reports can contain the logs from cached builds.

This worked by sheer luck for all packages for which the logs are
within the prefix (i.e. C++ packages), but did not for packages
that do not have an install step (e.g. Ruby packages).

This commit explicitly caches the package's log directory, and restores
it on pull